+++ maodevice/transmitter/__init__.py
@@ -1,10 +1,6 @@
 # -*- coding: utf-8 -*-
-from .md_20_m import Md20M
-from .pd_30_m import Pd30M
-from .lta_20_q import Lta20Q
+from .rfll_20_h import *
 from .model3390_awg import Model3390AWG
 
-del lta_20_q
-del md_20_m
-del pd_30_m
+del rfll_20_h
 del model3390_awg
